requirements.txt
Django==4.2
django_rest==0.8.7
djangorestframework==3.14.0
